Add files via upload
Exploring manual testing methodologies on my personal Rokomari project, focusing on ensuring quality and reliability through meticulous testing practices and continuous improvement. This project is solely for personal practice and experimentation with manual testing techniques. While it is not associated with the official Rokomari platform, it provides a hands-on opportunity to refine testing skills and explore different testing scenarios.